+++ main.c
@@ -2,8 +2,8 @@
 #include "HaMinhNghia.c"
 
 int main() {
-    int a, b, c;
-    a = 2; b = 4; c = 5;
+    float a, b, c;
+    a = 2.0; b = 4.0; c = 5.0;
     printf(add(a, b));
 
     return 0;
